Move iterators to a separate module
